uses 16 bit exponent entries

+++ src/data.h
@@ -44,7 +44,7 @@
 typedef int32_t ind_t;    /* index in hash table structure */
 typedef int32_t sdm_t;    /* short divmask for faster divisibility checks */
 typedef int32_t len_t;    /* length type for different structures */
-typedef int32_t exp_t;    /* exponent type */
+typedef int16_t exp_t;    /* exponent type */
 typedef int32_t val_t;    /* core values like hashes, coefficients, etc. */
 typedef val_t hl_t;       /* length of hash table */
 typedef int32_t deg_t;    /* (total) degree of polynomial */
